+++ wrench/grouper/kinetic/cooccurence.py
@@ -1,6 +1,7 @@
 from collections import defaultdict
 
 import community as cd
+import matplotlib.pyplot as plt
 import networkx as nx
 
 from .models import Cluster
@@ -46,6 +47,42 @@ def build_cooccurence_network(
         top = sorted(deg, key=deg.get, reverse=True)[:top_n]
         essential[f"cluster_{comm_id}"] = top
 
+    visualize_cooccurence_network(G, partition)
+
     return [
         Cluster(cluster_id=id, keywords=keywords) for id, keywords in essential.items()
     ]
+
+
+def visualize_cooccurence_network(G: nx.Graph, partition: dict):
+    """
+    Visualizes the cooccurence graph.
+
+    Displays the detected partitions in the graph.
+
+    Args:
+        G (nx.Graph): Graph containing nodes of keywords.
+        partition (dict): Partitioned nodes.
+    """
+    centrality = dict(G.degree(weight="weight"))
+
+    # Generate positions for all nodes
+    pos = nx.spring_layout(G, weight="weight", seed=42)
+
+    # Extract community IDs and centrality values
+    colors = [partition[node] for node in G.nodes()]
+    # Adjust node sizes: ensure a minimum size and scale by centrality
+    sizes = [max(100, centrality.get(node, 0) * 300) for node in G.nodes()]
+
+    # Draw nodes
+    nx.draw_networkx_nodes(G, pos, node_color=colors, node_size=sizes, cmap=plt.cm.Set3)
+
+    # Draw edges
+    nx.draw_networkx_edges(G, pos, alpha=0.5)
+
+    # Draw labels
+    nx.draw_networkx_labels(G, pos, font_size=6)
+
+    # Display the plot
+    plt.axis("off")
+    plt.show()
