+++ wrench/grouper/kinetic/prompts/embed_documents.txt
