+++ wrench/grouper/kinetic/prompts/embed_topics.txt
