+++ wrench/grouper/kinetic/prompts/generate_seed_topics.txt
