+++ wrench/grouper/kinetic/prompts/user_prompt.txt
